+++ js/const.js
@@ -1,0 +1,29 @@
+const MaxCountPhotos = 25;
+const COUNT_COMMETS = 5;
+const CountLikes = {
+    MIN: 15,
+    MAX: 200
+};
+const NAMES = [
+    'Анастасия',
+    'Мария',
+    'Александр',
+    'Артем',
+    'Юлия'
+];
+const MASSAGES = [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+const DISCRIPTION = [
+    'Любовь в каждом цветочке.',
+    'Радость моя!',
+    'Погода решила взгрустнуть.',
+    'Привет моя грусть, я теперь не боюсь.',
+];
+
+export {DISCRIPTION, MASSAGES, NAMES, CountLikes, COUNT_COMMETS, MaxCountPhotos}
